rework GameDataPlayer model, refactoring mock celery tasks/serializer (#489)
