+++ adaptive_hockey_federation/games/models.py
@@ -7,6 +7,7 @@ from django.db.models.functions import Now
 from django.utils.translation import gettext_lazy as _
 
 from games.constants import NumericalValues
+from main.models import Player
 
 
 class Game(models.Model):
@@ -148,3 +149,27 @@ class GamePlayer(models.Model):
     def __str__(self):
         """Метод, использующий поле name для строкового представления."""
         return f"{self.name} {self.last_name}"
+
+
+class GameDataPlayer(models.Model):
+    """Модель хранения JSON данных игроков для нарезки видео."""
+
+    player = models.ForeignKey(
+        Player,
+        on_delete=models.CASCADE,
+    )
+    game = models.ForeignKey(
+        Game,
+        on_delete=models.CASCADE,
+    )
+    data = models.JSONField(default=dict, verbose_name=_("Данные игры"))
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        verbose_name = _("JSON данные игрока для нарезки видео")
+        verbose_name_plural = _("JSON данные игроков для нарезки видео")
+        default_related_name = "game_data_player"
+
+    def __str__(self):
+        """Возвращает строку, содержащую имя игрока и дату создания."""
+        return f"{self.player.name} - {self.created_at}"
