+++ adaptive_hockey_federation/video_api/tasks.py
@@ -6,7 +6,7 @@ from celery.signals import task_success, worker_process_init
 from core.celery import app
 from service.a_hockey_requests import send_request_to_video_processing_service
 from service.video_processing import slicing_video_with_player_frames
-from main.models import GameDataPlayer
+from games.models import GameDataPlayer
 from .serializers import GameDataPlayerSerializer
 from celery import current_app
 from django.db import transaction
@@ -25,7 +25,7 @@ def get_player_video_frames(*args, **kwargs):
 def create_player_video(*args, **kwargs):
     # TODO моковая реализация, должно передаваться данные
     # необходимые для нарезки видео с игроком
-    time.sleep(30)
+    time.sleep(10)
     args = args[0]
     game_link = "https://disk.yandex.ru/i/JLh__1IbAfmK-Q"
     output_file = (
@@ -40,20 +40,22 @@ def bulk_create_gamedataplayer_objects(sender=None, **kwargs):
     # result - получает результат работы из sender
     result = kwargs.get("result")
     parse_data = json.loads(result)
+
     serializer = GameDataPlayerSerializer(data=parse_data)
     if serializer.is_valid():
         object_data = serializer.validated_data
-
-        game_id = object_data["game_id"]
+        # В тестовом запросе индекс у игры 0, изменить
+        # на любой другой индекс игры в бд
+        game = object_data["game_id"]
         tracking = object_data["tracking"]
         with transaction.atomic():
             for track in tracking:
-                player = track.get("player_id")
+                player = track["player_id"]
                 del track["player_id"]
-                track["game_id"] = game_id
                 # TODO возможно следует использовать bulk_create
                 GameDataPlayer.objects.update_or_create(
                     player=player,
+                    game=game,
                     data=json.dumps(track),
                 )
                 print(
@@ -65,11 +67,12 @@ def bulk_create_gamedataplayer_objects(sender=None, **kwargs):
                 # TODO в args передают аргументы
                 # нужные для нарезки видео с игроком
                 create_player_video.apply_async(
-                    args=["Обработка с низки приоритетом"],
+                    args=["Обработка с низким приоритетом"],
                     queue="slice_player_video_queue",
                     priority=255,
                 )
-    # Функция ничего не возвращает, в логах воркера она тоже не отслеживается.
+    print(serializer.errors)
+    # Задача ничего не возвращает.
     # Возможно из-за способа её вызова через worker_process_init
